% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_bigrams_network.R
\name{calc_bigrams_network}
\alias{calc_bigrams_network}
\title{Create and count bigrams}
\usage{
calc_bigrams_network(
  x,
  target_col_name,
  text_col_name,
  grouping_variables = NULL,
  filter_class = NULL,
  filter_main_group = NULL,
  bigrams_prop
)
}
\arguments{
\item{x}{A data frame with three columns: the column with the classes; the
column with the text; and the column(s) with the group(s).}

\item{target_col_name}{A string with the column name of the target variable.}

\item{text_col_name}{A string with the column name of the text variable.}

\item{grouping_variables}{A string or vector of strings with the column
name(s) (if any) of the grouping variable(s). Defaults to \code{NULL}.}

\item{filter_class}{A string or vector of strings with the name(s) of the
class(es) for which bigrams are to be created and counted. Defaults to
\code{NULL} (all classes).}

\item{filter_main_group}{A string with the name(s) of the organization(s)
for which to create and count bigrams. Defaults to \code{NULL} (all
organizations).}

\item{bigrams_prop}{A numeric in (0, 100] indicating the percentage of the
most frequent bigrams to keep.}
}
\value{
A data frame with three columns: first word of bigram;
second word of bigram; and bigram count.
}
\description{
For a given labeled text, create and calculate the most frequently occurring
bigrams for the class(es) and organization(s).
}
\note{
When supplying more than one organization and/or class, the returned
data frame will NOT separate the results for the different organizations
and/or classes, i.e. the function sees multiple organizations and/or
classes as one.
}
