% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_confusion_matrix.R
\name{calc_confusion_matrix}
\alias{calc_confusion_matrix}
\title{Title}
\usage{
calc_confusion_matrix(
  x,
  target_col_name,
  target_pred_col_name,
  grouping_variable = NULL
)
}
\arguments{
\item{grouping_variable}{}
}
\value{

}
\description{
Title
}
