% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_predict_unlabelled_text.R
\name{calc_predict_unlabelled_text}
\alias{calc_predict_unlabelled_text}
\title{Predict unlabelled text using a fitted Scikit-learn pipeline}
\usage{
calc_predict_unlabelled_text(
  x,
  python_setup = TRUE,
  sys_setenv,
  which_python,
  which_venv,
  venv_name,
  text_col_name,
  pipe_path,
  preds_column = NULL,
  column_names = "all_cols"
)
}
\arguments{
\item{x}{A data frame with the text data to predict classes for.}

\item{python_setup}{A \code{logical} whether to set up the \code{Python} version,
virtual environment etc. that can be controlled with arguments
\code{sys_setenv}, \code{which_python}, \code{which_venv} and \code{venv_name}. These
arguments will be ignored when \code{python_setup} is \code{FALSE}. The purpose of
\code{python_setup} is that users may wish to control the \code{Python} parameters
outside the actual function, for the session in general.}

\item{sys_setenv}{A string in the form "path_to_python/python.exe",
indicating which Python to use (e.g. from a virtual environment).}

\item{which_python}{Same as \code{sys_setenv}.}

\item{which_venv}{A string that can be "conda", "miniconda" or "python".}

\item{venv_name}{String. The name of the virtual environment.}

\item{text_col_name}{A string with the column name of the text variable.}

\item{pipe_path}{A string in the form "path_to_fitted_pipeline/pipeline.sav",
where "pipeline" is the name of the SAV file with the fitted
\code{Scikit-learn} pipeline.}

\item{preds_column}{A string with the user-specified name of the column that
will have the predictions. If \code{NULL} (default), then the name will be
\code{paste0(text_col_name, "_preds")}.}

\item{column_names}{A vector of strings with the names of the columns of the
supplied data frame (incl. \code{text_col_name}) to be added to the returned
data frame. If "preds_only", then the only column in the returned data
frame will be \code{preds_column}. Defaults to "all_cols".}
}
\value{

}
\description{
Predict unlabelled text using a fitted Scikit-learn pipeline
}
