% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bigrams_network.R
\name{get_bigrams_network}
\alias{get_bigrams_network}
\title{Title}
\usage{
get_bigrams_network(
  x,
  target_col_name,
  filter_class,
  filter_organization,
  bigrams_prop
)
}
\arguments{
\item{ngrams_type}{}
}
\value{

}
\description{
Title
}
