% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bing_word_counts.R
\name{get_bing_word_counts}
\alias{get_bing_word_counts}
\title{Title}
\usage{
get_bing_word_counts(x, target_col_name, filter_organization, filter_class)
}
\arguments{
\item{filter_class}{}
}
\value{

}
\description{
Title
}
