% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dictionary.R
\name{get_dictionary}
\alias{get_dictionary}
\title{Check for sentiment dictionaries}
\usage{
get_dictionary(dictionary)
}
\arguments{
\item{dictionary}{string. One of "afinn", "nrc", or "bing", indicating
the dictionary to be loaded}
}
\value{
a dataframe containing the requested sentiment dictionary
}
\description{
On a server, the sentiment dictionaries have to be manually uploaded. On a
local machine they can be loaded using {tidytext}. Check to see if they
are already loaded (by a previous function run), if they are on disk
(which would indicate the code is on a server), or just load using
{tidytext}
}
