% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_net_sentiment_long_nrc.R
\name{get_net_sentiment_long_nrc}
\alias{get_net_sentiment_long_nrc}
\title{Title}
\usage{
get_net_sentiment_long_nrc(
  net_sentiment_wide_nrc,
  sorting_sentiments = "anger",
  num_of_facets = 60
)
}
\value{

}
\description{
Title
}
