% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_net_sentiment_per_tag.R
\name{get_net_sentiment_per_tag}
\alias{get_net_sentiment_per_tag}
\title{Title}
\usage{
get_net_sentiment_per_tag(x, target_col_name, filter_organization)
}
\arguments{
\item{filter_organization}{}
}
\value{

}
\description{
Title
}
