% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_net_sentiment_wide_nrc.R
\name{get_net_sentiment_wide_nrc}
\alias{get_net_sentiment_wide_nrc}
\title{Title}
\usage{
get_net_sentiment_wide_nrc(
  x,
  class_col_name,
  org_col_name,
  filter_class,
  filter_organization
)
}
\arguments{
\item{filter_organization}{}
}
\value{

}
\description{
Title
}
