% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tfidf_ngrams.R
\name{get_tfidf_ngrams}
\alias{get_tfidf_ngrams}
\title{Title}
\usage{
get_tfidf_ngrams(
  x,
  target_col_name,
  filter_class,
  filter_organization,
  ngrams_type
)
}
\arguments{
\item{ngrams_type}{}
}
\value{

}
\description{
Title
}
