% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_net_sentiment_per_tag.R
\name{plot_net_sentiment_per_tag}
\alias{plot_net_sentiment_per_tag}
\title{Title}
\usage{
plot_net_sentiment_per_tag(net_sentiment_all_dicts, target_col_name)
}
\arguments{
\item{target_col_name}{}
}
\value{

}
\description{
Title
}
