% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tfidf_ngrams.R
\name{plot_tfidf_ngrams}
\alias{plot_tfidf_ngrams}
\title{Title}
\usage{
plot_tfidf_ngrams(tfidf_ngrams, ngrams_type, filter_class)
}
\arguments{
\item{tfidf_ngrams}{}
}
\value{

}
\description{
Title
}
