% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_best_estimators.R
\name{prep_best_estimators}
\alias{prep_best_estimators}
\title{Prepare table with the best (hyper)parameter tunings for each learner}
\usage{
prep_best_estimators(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Internal function
}
