% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_colnames_and_filters.R
\name{prep_colnames_and_filters}
\alias{prep_colnames_and_filters}
\title{Prepare column names and filtering vectors for use in tables}
\usage{
prep_colnames_and_filters(
  x = NULL,
  grouping_variables = NULL,
  target_col_name,
  filter_class = NULL,
  filter_main_group = NULL,
  column_names = NULL
)
}
\arguments{
\item{column_names}{}
}
\value{

}
\description{
Internal function
}
