% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_python.R
\name{prep_python}
\alias{prep_python}
\title{Title}
\usage{
prep_python(sys_setenv, which_python, which_venv, venv_name, text_col_name)
}
\arguments{
\item{text_col_name}{}
}
\value{

}
\description{
Title
}
\examples{

}
