% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_sentiments_nrc.R
\name{prep_sentiments_nrc}
\alias{prep_sentiments_nrc}
\title{Pulls NRC Sentiments}
\usage{
prep_sentiments_nrc()
}
\value{
Character vector of all NRC sentiments
}
\description{
Pulls NRC Sentiments
}
