% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_tidy_feedback.R
\name{prep_tidy_feedback}
\alias{prep_tidy_feedback}
\title{Unnest tokens for each label in a labeled text}
\usage{
prep_tidy_feedback(
  x,
  target_col_name,
  text_col_name,
  grouping_variables = NULL
)
}
\arguments{
\item{x}{A data frame with three columns: the column with the classes; the
column with the text; and the column(s) with the group(s).}

\item{target_col_name}{A string with the column name of the target variable.}

\item{text_col_name}{A string with the column name of the text variable.}

\item{grouping_variables}{A string or vector of strings with the column
name(s) (if any) of the grouping variable(s). Defaults to \code{NULL}.}
}
\value{

}
\description{
Creates table with tokens and other columns (e.g. labels, groups)
passed by the user.
}
