% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_tidy_text.R
\name{prep_tidy_text}
\alias{prep_tidy_text}
\title{Unnest tokens for each label in a labelled text}
\usage{
prep_tidy_text(x, target_col_name = NULL, text_col_name)
}
\arguments{
\item{x}{x A data frame with one or more columns: the column with the classes
(if \code{target_col_name} is not \code{NULL}); and the column with the text. Any
other columns will be ignored.}

\item{target_col_name}{A string with the column name of the target variable.
Defaults to \code{NULL}.}

\item{text_col_name}{A string with the column name of the text variable.}
}
\value{
A data frame with two or three columns: classes (if \code{target_col_name}
is not \code{NULL}); line numbers; and tokens.
}
\description{
Creates table with tokens for each class (if any).
}
\examples{
#' library(experienceAnalysis)
books <- janeaustenr::austen_books() # Jane Austen books
emma <- paste(books[books$book == "Emma", ], collapse = " ") # String with whole book
pp <- paste(books[books$book == "Pride & Prejudice", ], collapse = " ") # String with whole book

# Make data frame with books Emma and Pride & Prejudice
x <- data.frame(
  text = c(emma, pp),
  book = c("Emma", "Pride & Prejudice")
)

# Tokens for both books, without the `target_col_name` column
prep_tidy_text(x, target_col_name = NULL, text_col_name = "text") \%>\%
    head()

# Tokens for both books, with the `target_col_name` column this time
prep_tidy_text(x, target_col_name = "book", text_col_name = "text") \%>\%
    split(.$book) \%>\%
    purrr::map(~ head(.))

# Tokens for Pride & Prejudice
prep_tidy_text(x, target_col_name = "book", text_col_name = "text") \%>\%
    dplyr::filter(book == "Pride & Prejudice") \%>\%
    head()
}
