% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_class_filter.R
\name{tidy_class_filter}
\alias{tidy_class_filter}
\title{Filter data frame by the supplied string vector}
\usage{
tidy_class_filter(x, filter_class = NULL)
}
\arguments{
\item{filter_class}{}
}
\value{

}
\description{
For use inside \code{dplyr::filter()} operations in different \code{experienceAnalysis}
functions.
}
