% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_net_sentiment_nrc.R
\name{tidy_net_sentiment_nrc}
\alias{tidy_net_sentiment_nrc}
\title{Title}
\usage{
tidy_net_sentiment_nrc(
  net_sentiment_wide_nrc,
  sorting_sentiments = "anger",
  num_of_facets = 60
)
}
\value{

}
\description{
Title
}
