% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_accuracy_per_class.R
\name{calc_accuracy_per_class}
\alias{calc_accuracy_per_class}
\title{Title}
\usage{
calc_accuracy_per_class(
  x,
  target_col_name,
  target_pred_col_name,
  grouping_variable = NULL
)
}
\arguments{
\item{grouping_variable}{}
}
\value{

}
\description{
Title
}
