% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_bigrams_network.R
\name{calc_bigrams_network}
\alias{calc_bigrams_network}
\title{Title}
\usage{
calc_bigrams_network(
  x,
  target_col_name,
  filter_class,
  filter_organization,
  bigrams_prop
)
}
\arguments{
\item{ngrams_type}{}
}
\value{

}
\description{
Title
}
