% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_bing_word_counts.R
\name{calc_bing_word_counts}
\alias{calc_bing_word_counts}
\title{Title}
\usage{
calc_bing_word_counts(x, target_col_name, filter_organization, filter_class)
}
\arguments{
\item{filter_class}{}
}
\value{

}
\description{
Title
}
