% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_bing_word_counts.R
\name{calc_bing_word_counts}
\alias{calc_bing_word_counts}
\title{Counts of words with a positive or negative sentiment}
\usage{
calc_bing_word_counts(
  x,
  target_col_name = NULL,
  text_col_name,
  filter_class = NULL
)
}
\arguments{
\item{x}{A data frame with one or more columns: the column with the classes
(if \code{target_col_name} is not \code{NULL}); and the column with the text. Any
other columns will be ignored.}

\item{target_col_name}{A string with the column name of the target variable.
Defaults to \code{NULL}.}

\item{text_col_name}{A string with the column name of the text variable.}

\item{filter_class}{A string or vector of strings with the name(s) of the
class(es) for which to count the words. Defaults to
\code{NULL} (all rows).}
}
\value{
A data frame with three columns: word; sentiment ("positive" or
"negative"- see Hu & Liu, 2004); and count.
}
\description{
Count the number of times a word with a positive or negative sentiment occurs
in a given text.
}
\note{
When supplying more than one class in \code{filter_class}, the returned data
frame will NOT separate the results for the different classes. If
separation is desired, then do something like this:\preformatted{# Assuming that the class and text text columns are called "label" and
# "feedback" respectively
x \%>\%
    split(.$label) \%>\%
    purrr::map(
        ~ calc_bing_word_counts(., target_col_name = NULL,
                               text_col_name = "feedback",
                               filter_class = NULL)
    )
}
}
\references{
Hu M. & Liu B. (2004). Mining and summarizing customer
reviews. Proceedings of the ACM SIGKDD International Conference on
Knowledge Discovery & Data Mining (KDD-2004), Seattle, Washington, USA,
Aug 22-25, 2004.
}
