% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_confusion_matrix.R
\name{calc_confusion_matrix}
\alias{calc_confusion_matrix}
\title{Calculate the confusion matrix}
\usage{
calc_confusion_matrix(x, target_col_name, target_pred_col_name, ...)
}
\arguments{
\item{x}{A data frame with two columns: the column with the actual classes;
and the column with the predicted classes. Any other columns will be
ignored.}

\item{target_col_name}{A string with the column name of the target variable.}

\item{target_pred_col_name}{A string with the column name of the predictions
for the target variable.}

\item{...}{Further arguments passed from other methods.}
}
\value{
An object with class \code{conf_mat()} (see \verb{conf_mat\{yardstick\}}).
}
\description{
Calculates the confusion matrix of observed and predicted classes, after
converting them into factors first (not currently done by
\verb{conf_mat\{yardstick\}}).
}
\examples{
library(experienceAnalysis)
mtcars \%>\%
  dplyr::mutate(carb_pred = sample(carb, size = nrow(.))) \%>\%  # Mock predictions column
  calc_accuracy_per_class(
    target_col_name = "carb",
    target_pred_col_name = "carb_pred"
  )

# Custom column names
mtcars \%>\%
  dplyr::mutate(carb_pred = sample(carb, size = nrow(.))) \%>\%  # Mock predictions column
  calc_confusion_matrix(
    target_col_name = "carb",
    target_pred_col_name = "carb_pred"
  )
}
