% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_net_sentiment_nrc.R
\name{calc_net_sentiment_nrc}
\alias{calc_net_sentiment_nrc}
\title{Title}
\usage{
calc_net_sentiment_nrc(
  x,
  class_col_name,
  org_col_name,
  filter_class,
  filter_organization
)
}
\arguments{
\item{filter_organization}{}
}
\value{

}
\description{
Title
}
