% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_predict_unlabelled_text.R
\name{calc_predict_unlabelled_text}
\alias{calc_predict_unlabelled_text}
\title{Predict unlabelled text using a fitted Scikit-learn pipeline}
\usage{
calc_predict_unlabelled_text(
  x,
  file_path = NULL,
  predictor,
  pipe_path,
  preds_column = NULL,
  column_names = NULL
)
}
\arguments{
\item{x}{A data frame with the text data to predict classes for. If
\code{file_path} (see below) is not \code{NULL}, then \code{x} should be \code{NULL}.}

\item{file_path}{A string in the form "path_to_data/filename.csv", where
"filename" is the name of the CSV file with the data. If \code{NULL}, the
function uses x.}

\item{predictor}{A string with the name of the text column.}

\item{pipe_path}{A string in the form "path_to_fitted_pipeline/pipeline.sav,"
where "pipeline" is the name of the SAV file with the fitted
\code{Scikit-learn} pipeline.}

\item{preds_column}{A string with the user-specified name of the column that
will have the predictions. If \code{NULL} (default), then the name will be
\code{paste0(predictor, "_preds")}.}

\item{column_names}{A vector of strings with the names of the columns of the
supplied data frame (incl. \code{predictor}) to be added to the returned data
frame. If \code{NULL}, then the only column in the returned data frame will be
\code{preds_column.}}

\item{sys_setenv}{A string in the form "path_to_python/python.exe",
indicating which Python to use (e.g. from a virtual environment).}

\item{which_python}{Same as \code{sys_setenv}.}

\item{which_venv}{A string that can be "conda", "miniconda" or "python".}

\item{venv_name}{String. The name of the virtual environment.}
}
\value{

}
\description{
For internal use only!
}
