% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_sentiment_indicators.R
\name{calc_sentiment_indicators}
\alias{calc_sentiment_indicators}
\title{Calculate TextBlob sentiment indicators}
\usage{
calc_sentiment_indicators(
  x,
  sys_setenv,
  which_python,
  which_venv,
  venv_name,
  make_table = TRUE,
  text_col_name
)
}
\arguments{
\item{make_table}{}
}
\value{

}
\description{
Calculate TextBlob sentiment indicators
}
