% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_confusion_matrix.R
\name{plot_confusion_matrix}
\alias{plot_confusion_matrix}
\title{Title}
\usage{
plot_confusion_matrix(
  x,
  target_col_name,
  target_pred_col_name,
  type = "heatmap"
)
}
\arguments{
\item{type}{}
}
\value{

}
\description{
Title
}
