% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_net_sentiment_long_nrc.R
\name{plot_net_sentiment_long_nrc}
\alias{plot_net_sentiment_long_nrc}
\title{Title}
\usage{
plot_net_sentiment_long_nrc(net_sentiment_long_nrc)
}
\arguments{
\item{net_sentiment_long_nrc}{}
}
\value{

}
\description{
Title
}
