% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_tidy_feedback.R
\name{prep_tidy_feedback}
\alias{prep_tidy_feedback}
\title{Unnest tokens for each label in a labelled text}
\usage{
prep_tidy_feedback(x, target_col_name = NULL, text_col_name)
}
\arguments{
\item{x}{x A data frame with one or more columns: the column with the classes
(if \code{target_col_name} is not \code{NULL}); and the column with the text. Any
other columns will be ignored.}

\item{target_col_name}{A string with the column name of the target variable.
Defaults to \code{NULL}.}

\item{text_col_name}{A string with the column name of the text variable.}
}
\value{
A data frame with two or three columns: classes (if \code{target_col_name}
is not \code{NULL}); line numbers; and tokens.
}
\description{
Creates table with tokens for each class (if any).
}
